refactor(frontend): refactor ImageEffector
